+++ shared/session/session.go
@@ -1,12 +1,74 @@
 package session
 
-import "net/http"
+import (
+	"github.com/JohanAanesen/NTNU-Bachelor-Management-System-For-CS-Assignments/db"
+	"github.com/JohanAanesen/NTNU-Bachelor-Management-System-For-CS-Assignments/internal/model"
+	"log"
+	"net/http"
+)
 
-type Session struct {
-	Values map[string]interface{}
+//IsTeacher returns if user is a teacher or not
+func IsTeacher(r *http.Request) bool {
+	//check if user is already logged in
+	user := GetUserFromSession(r)
+
+	//check that user is a teacher
+	if !user.Teacher { //not a teacher or logged in
+		return false
+	}
+
+	return IsLoggedIn(r)
 }
 
-func Instance(r *http.Request) *Session {
-	return &Session{}
+//IsLoggedIn returns if user is authenticated or not
+func IsLoggedIn(r *http.Request) bool {
+	//get user from session
+	user := GetUserFromSession(r)
+
+	//check that user is a teacher
+	if !user.Authenticated { //not logged in
+		return false
+	}
+
+	return true
 }
 
+//GetUserFromSession returns user object stored in session
+func GetUserFromSession(r *http.Request) model.User {
+	session, err := db.CookieStore.Get(r, "login-session") //get session
+	if err != nil {
+		log.Println(err)
+		return model.User{Authenticated: false}
+	}
+
+	val := session.Values["user"]
+	var user = model.User{}
+	user, ok := val.(model.User)
+	if !ok {
+		return model.User{Authenticated: false}
+	}
+	return user
+}
+
+//SaveUserToSession saves user object to sessionstore
+func SaveUserToSession(user model.User, w http.ResponseWriter, r *http.Request) bool {
+	session, err := db.CookieStore.Get(r, "login-session") //get session
+	if err != nil {
+		log.Println(err)
+		return false
+	}
+
+	session.Values["user"] = user
+
+	err = session.Save(r, w) //save session changes
+
+	if err != nil {
+		//todo log this event
+		log.Fatal(err)
+
+		//redirect somewhere
+		return false
+	}
+
+	return true
+}
